Dashboards: Widen v2 panel ID types from uint16 to int32 & >=0
uint16 caps panel IDs at 65,535. Fine for UI-created dashboards,
but dashboards-as-code (Terraform, jsonnet, k8s controllers) can
emit timestamps or arbitrary numerics as IDs. The v1 -> v2alpha1
converter then silently drops those.

uint16 also leaves a contract mismatch: CUE uint16 emits
format: int32 in OpenAPI, so the wire spec advertises int32 range
while the Go decoder rejects anything > 65535. Same class of
loose-contract bug as the original int64-with-string issue, just
inverted.

int32 & >=0 emits the same OpenAPI shape (format: int32, no
string coercion), widens the range to ~2.1B, and aligns the Go
runtime exactly with the wire. Migration silent-drop path tightened
to MaxInt32 -- unreachable in practice.

Per @kristinademeshchik's review on the original PR.

+++ apps/dashboard/pkg/migration/conversion/conversion_data_loss_detection_test.go
@@ -2285,7 +2285,7 @@ func createV2DashboardWithSimpleTabs() *dashv2alpha1.Dashboard {
 			PanelKind: &dashv2alpha1.DashboardPanelKind{
 				Kind: "Panel",
 				Spec: dashv2alpha1.DashboardPanelSpec{
-					Id:        uint16(i),
+					Id:        int32(i),
 					Title:     fmt.Sprintf("Panel %d - Tab %d", i, tabNum),
 					VizConfig: dashv2alpha1.DashboardVizConfigKind{Kind: "timeseries"},
 					Data: dashv2alpha1.DashboardQueryGroupKind{
@@ -2360,7 +2360,7 @@ func createV2DashboardWithRowsCollapsed() *dashv2alpha1.Dashboard {
 			PanelKind: &dashv2alpha1.DashboardPanelKind{
 				Kind: "Panel",
 				Spec: dashv2alpha1.DashboardPanelSpec{
-					Id:        uint16(i),
+					Id:        int32(i),
 					Title:     fmt.Sprintf("Panel %d - Row %d (collapsed)", i, rowNum),
 					VizConfig: dashv2alpha1.DashboardVizConfigKind{Kind: "timeseries"},
 					Data: dashv2alpha1.DashboardQueryGroupKind{
@@ -2438,7 +2438,7 @@ func createV2DashboardWithRowsExpanded() *dashv2alpha1.Dashboard {
 			PanelKind: &dashv2alpha1.DashboardPanelKind{
 				Kind: "Panel",
 				Spec: dashv2alpha1.DashboardPanelSpec{
-					Id:        uint16(i),
+					Id:        int32(i),
 					Title:     fmt.Sprintf("Panel %d - Row %d (expanded)", i, rowNum),
 					VizConfig: dashv2alpha1.DashboardVizConfigKind{Kind: "timeseries"},
 					Data: dashv2alpha1.DashboardQueryGroupKind{
@@ -2881,7 +2881,7 @@ func createV2ComplexNestedDashboard() *dashv2alpha1.Dashboard {
 			PanelKind: &dashv2alpha1.DashboardPanelKind{
 				Kind: "Panel",
 				Spec: dashv2alpha1.DashboardPanelSpec{
-					Id:        uint16(i),
+					Id:        int32(i),
 					Title:     fmt.Sprintf("Panel %d", i),
 					VizConfig: dashv2alpha1.DashboardVizConfigKind{Kind: "timeseries"},
 					Data: dashv2alpha1.DashboardQueryGroupKind{
